:children_crossing: 扩展商店 - 如果扩展已经安装则不再进行重复下载和安装

+++ SRC/Host/AuraEcho/Services/MarketPluginInstallTask.cs
@@ -1,8 +1,10 @@
 using System;
 using System.IO;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using AuraEcho.Core.Contracts;
+using AuraEcho.Core.Data.Entities;
 using AuraEcho.Core.Events;
 using AuraEcho.Core.Models;
 using AuraEcho.Core.Tools;
@@ -66,7 +68,6 @@ public class MarketPluginInstallTask : BindableBase, ITransferTask
     public async Task Start()
     {
         if (_inProgress) return;
-
         _cts = new CancellationTokenSource();
         try
         {
@@ -81,9 +82,12 @@ public class MarketPluginInstallTask : BindableBase, ITransferTask
             string packageFilePath = await DownloadAsync(_cts.Token);
 
             Status = MarketPluginInstallStatus.Installing;
-            await InstallAsync(packageFilePath, _cts.Token);
+            var localPlugin = await InstallAsync(packageFilePath, _cts.Token);
+            var userPlugin = await RegisterUserPluginAsync(localPlugin);
+            await LoadPluginAsync(userPlugin, _cts.Token);
 
             Status = MarketPluginInstallStatus.Completed;
+            _eventAggregator.GetEvent<PluginInstalledEvent>().Publish(userPlugin);
         }
         catch (OperationCanceledException)
         {
@@ -97,6 +101,23 @@ public class MarketPluginInstallTask : BindableBase, ITransferTask
         {
             _inProgress = false;
         }
+    }
+
+    private async Task<(bool, LocalPluginModel?)> IsInstalled(MarketPlugin mp)
+    {
+        var latestVersionPackInfo = await _remotePluginRepository.GetLatestAsync(mp.PluginInfo.Id);
+        Version? marketPluginLatestVersion =
+            latestVersionPackInfo is null
+            ? null
+            : Version.Parse(latestVersionPackInfo.Version);
+
+        var localPlugins = await _localPluginRepository.GetLocalPluginsAsync();
+        LocalPluginModel? installedPlugin = 
+            localPlugins.FirstOrDefault(lp => 
+                lp.Id == mp.PluginInfo.Id && 
+                Version.Parse(lp.Manifest.Version!) == marketPluginLatestVersion);
+
+        return (installedPlugin is not null, installedPlugin);
     }
 
     public void Cancel()
@@ -147,6 +168,15 @@ public class MarketPluginInstallTask : BindableBase, ITransferTask
     // 下载
     private async Task<string> DownloadAsync(CancellationToken token)
     {
+        if (await IsInstalled(_plugin) is (true, _))
+        {
+            await Task.Delay(TimeSpan.FromSeconds(0.5), token);
+            Progress = 50;
+            await Task.Delay(TimeSpan.FromSeconds(0.5), token);
+            Progress = 100;
+            return String.Empty;
+        }
+
         var pluginInstallerFilePath = Path.Combine(ApplicationPaths.Temp, $"{_plugin.PluginInfo.Id}.plix");
         Progress<double> progressHandler = new Progress<double>(p => Progress = p);
         await Task.Delay(TimeSpan.FromSeconds(0.5));
@@ -164,21 +194,49 @@ public class MarketPluginInstallTask : BindableBase, ITransferTask
 
         return pluginInstallerFilePath;
     }
-    // 安装
-    private async Task InstallAsync(string packageFilePath, CancellationToken token)
+    /// <summary>
+    /// 安装
+    /// </summary>
+    /// <param name="packageFilePath"></param>
+    /// <param name="token"></param>
+    /// <returns></returns>
+    /// <exception cref="Exception"></exception>
+    private async Task<LocalPluginModel> InstallAsync(string packageFilePath, CancellationToken token)
     {
+        if (await IsInstalled(_plugin) is (true, var installedPlugin))
+        {
+            await Task.Delay(TimeSpan.FromSeconds(0.5), token);
+            return installedPlugin!;
+        }
         Task<LocalPluginModel> installlTask = _pluginInstallService.InstallAsync(packageFilePath);
         await Task.WhenAll(installlTask, Task.Delay(TimeSpan.FromSeconds(0.5), token));
         var localPlugin = await installlTask;
         File.Delete(packageFilePath);
-        if (localPlugin is null) throw new Exception("Plugin install failed");
-        var addUserPlugin = await _localPluginRepository.AddUserPluginAsync(_clientSession.CurrentUser.Id, localPlugin.Id);
 
-        var loadPluginTask = _pluginManager.LoadPluginAsync(addUserPlugin);
+        return localPlugin ?? throw new Exception("Plugin install failed");
+    }
+
+    /// <summary>
+    /// 注册用户插件关系
+    /// </summary>
+    /// <returns></returns>
+    private async Task<UserPluginModel> RegisterUserPluginAsync(LocalPluginModel plugin)
+    {
+        var addUserPlugin = await _localPluginRepository.AddUserPluginAsync(_clientSession.CurrentUser.Id, plugin.Id);
+        return addUserPlugin;
+    }
+
+    /// <summary>
+    /// 加载插件
+    /// </summary>
+    /// <returns></returns>
+    private async Task LoadPluginAsync(UserPluginModel up, CancellationToken token)
+    {
+        var loadPluginTask = _pluginManager.LoadPluginAsync(up);
         await Task.WhenAll(loadPluginTask, Task.Delay(TimeSpan.FromSeconds(0.5), token));
         var installResult = await loadPluginTask;
-        if (!installResult) throw new Exception("Plugin load failed");
 
-        _eventAggregator.GetEvent<PluginInstalledEvent>().Publish(addUserPlugin);
+        if (!installResult) 
+            throw new Exception("Plugin load failed");
     }
 }
